refactor: move and rename the base counter class in counter.py module
